[x64] Factor out control code
